#18963 Minimize compiler warnings

+++ src/GLViewer/GLViewer_Object.cxx
@@ -196,7 +196,7 @@ QByteArray GLViewer_Object::getByteCopy()
 
     aCurPos += aGLText.size();
     aPointer = (char*)&myOwner;
-    for( i = 0; i < sizeof( SUIT_DataOwner* ); i++, aPointer++ )
+    for( i = 0; i < (int)sizeof( SUIT_DataOwner* ); i++, aPointer++ )
         aResult[ aCurPos + i ] = *aPointer;
 
     return aResult;
