+++ src/Session/Session_Session_i.cxx
@@ -349,16 +349,16 @@ void SALOME_Session_i::deleteContainersinNS()
 // destroy of all containers and modules
   _NS->Change_Directory("/Containers");
   std::vector<std::string> machines = _NS->list_subdirs();
-  for(int i=0;i<machines.size();i++){
+  for(int i=0;i<(int)machines.size();i++){
     _NS->Change_Directory(machines[i].c_str());
     std::vector<std::string> toto = _NS->list_directory();
-    for(int j=0;j<toto.size();j++)
+    for(int j=0;j<(int)toto.size();j++)
       _NS->Destroy_Name(toto[j].c_str());
     std::vector<std::string> containers = _NS->list_subdirs();
-    for(int j=0;j<containers.size();j++){
+    for(int j=0;j<(int)containers.size();j++){
       _NS->Change_Directory(containers[j].c_str());
       std::vector<std::string> modules = _NS->list_directory();
-      for(int k=0;k<modules.size();k++)
+      for(int k=0;k<(int)modules.size();k++)
         _NS->Destroy_Name(modules[k].c_str());
       _NS->Change_Directory("/Containers");
       _NS->Change_Directory(machines[i].c_str());
